The type of monetary parameters in the database has been changed to decimal for accuracy, and all add methods now return an ID

+++ My_money/Model/BudgetCategory.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public float? Plan { get; set; }
-        public float? Spend { get; set; }
+        public decimal? Plan { get; set; }
+        public decimal? Spend { get; set; }
     }
 }
